Move IM tests into top level tree too
