+++ nettests/telegram.go
@@ -1,15 +1,11 @@
-package im
-
-import (
-	"github.com/ooni/probe-cli/nettests"
-)
+package nettests
 
 // Telegram test implementation
 type Telegram struct {
 }
 
 // Run starts the test
-func (h Telegram) Run(ctl *nettests.Controller) error {
+func (h Telegram) Run(ctl *Controller) error {
 	builder, err := ctl.Ctx.Session.NewExperimentBuilder(
 		"telegram",
 	)
